requirements.txt
altgraph==0.16.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
future==0.17.1
get==2019.3.22
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
macholib==1.11
PyJWT==1.7.1
MarkupSafe==1.1.1
pefile==2019.4.14
post==2019.3.22
public==2019.3.22
py4j==0.10.7
pywin32-ctypes==0.2.0
query-string==2019.3.22
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.1
